tests: Refactor a utility code
